stockage_service/pyproject.toml
[tool.poetry]
name = "stockage-service"
version = "0.1.0"
description = ""
authors = ["Jilvo <jmillot41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
kink = "^0.8.0"
sphinx = "^7.4.7"
sphinx-autodoc-typehints = "^2.2.3"
pymongo = {extras = ["srv"], version = "^4.8.0"}
sphinx-wagtail-theme = "^6.3.0"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
